implement camera calibration software
this application is separate from the rest of the project as it only
needs to run once per camera setup. the only shared piece of code is
`StereoCameraInfo` as it's the data format used for data exchange
between the two (through a stored file).

the project leverages [OpenCV's camera calibration module].

note that the flags used for the `cv::stereoCalibrate` function in
`RunCalibration` are taken 1:1 from the lighthouse project as we're
currently working with the test images from there. it's quite likely
that these should be changed (potentially just removed / set to the
default) once we start working with our own images.

[OpenCV's camera calibration module]: https://docs.opencv.org/4.10.0/d9/d0c/group__calib3d.html

closes #4