add image rectification
also re-format the files as they had the wrong indentation. sorry for
mixing this in one commit :(

this adds the rectification logic without actually calling it as this
hasn't been implemented yet. that will come as a follow-up. i've just
added it here to already show how that'll look like.