implement initial version of the static image viewer
this supports either running it with one set of images and a config file
or with a folder. In the latter case the images from the folder are
being processed in order and shown. this latter case is not supported
with GCCs libstdc++ at the moment due to the same bug already affecting
the calibration software. finding a workaround here would require more
code and is not strictly needed.

from now on the actual library API returns a value instead of erroring
out, however it does not yet do anything beyond rectifying the images.