manually add fix for GLIBC++ bug 117560
this is a 1:1 copy of the fix from GLIBC++ and it is only being
applied on GLIBC++ versions which are known to be affected (presuming
that no new GLIBC++ releases will be created without the fix). the
latest release (affected) is 14.2.0, the fix should be in 14.3.0.

this way we don't need a workaround for the calibration tool and can use
the folder functionality of the static image viewer also with GCC.