calibrate: only use images if both have same amount of points
otherwise the processing will crash when calling `cv::stereoCalibrate`
as `object_points_left` will not contain the same amount of entries as
`image_points_right` (note that `object_points_right` remains unused
since OpenCV expects both images to contain the exact same object
points).

this is still suboptimal since it _could_ be that both cameras see
different markers but the same amount. however, this is an unlikely case
and doesn't seem to have happened in our calibration so it's "good
enough" for now.