calibration v1: fix charuco board size & calibration data
turns out that i still had the wrong marker & square size in the config
file, thus the resulting calibration had the wrong sizes.
while this has no impact on rectification it is relevant for the
calculation of sizes based on this.

that this is now correct is visible in the fact that the translation
matrix shows 75mm y-distance between the cameras, which is what we also
measured physically.