implement distance calculation
this calculates a disparity map and uses it to calculate the 3D
reprojected points for given 2D points. based on these 3D points the
distance between them and the camera as well as the distance between the
two points can be found using simple geometry.

there's an open TODO in this: currently the two points are hardcoded.
this will be resolved in a follow-up commit which will make this
user-selectable in the GUI.

the test `RectifyImages` has been removed for the moment since it anyway
didn't test the behaviour. it can be re-added later.

resolves #8