implement disparity map calculation using ORB
this makes use of ORB to calculate matching points in the two images.
since the user is allowed to select points in the GUI we'll have to find
the closest points from ORB. this may or may not be anywhere close to
what the user had selected.

the code has been slightly modified to allow selecting the algorithm to
be used, this way we can easily switch between this algorithm and the
previous implementation (with the previous remaining the default as it
is better suited for our use-case).