image viewer: refactor
there's no need to pass the `StereoVision` instance every time, instead
we can just pass it to the constructor.

this also allows us to access the settings (which we now expose) and
check the algorithm - not all algorithms modify the input image and thus
we can paint the selected points ourselves in the result.