add 'live' folder for live calibration
this way the folder is always present but none of the files will ever be
added to git (just copy the content over to a new folder if you ran the
calibration and want to add it).
while you still need to create the `config.yml` you don't need to also
create the folders first.

this is just relevant if you're using the launch configuration which
uses `live` as the default folder.