calibration: also make use of partial matches
this is a big improvement over the existing situation: so far only
images where both cameras matched the same amount of points were used.
this even contained a small corner case which didn't work: if they
matched the same amount but not the same points the code would fail.

now we only pick the points which have been found in both images.